pyproject.toml
[tool.poetry]
name = "patternanalysis-2023"
version = "0.1.0"
description = ""
authors = ["Lars Moan <lars.ostberg.moan@gmail.com>"]
readme = "README.md"
packages = [{include = "patternanalysis_2023"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = ">=1.18.5,<1.24.0"
opencv-python = ">=4.1.1"
torch = ">=1.7.0,<1.12.0 || >1.12.0"
pyyaml = ">=5.3.1"
tensorboard = ">=2.4.1"
matplotlib = ">=3.2.2"
requests = ">=2.23.0"
scipy = ">=1.4.1"
torchvision = ">=0.8.1,<0.13.0 || >0.13.0"
tqdm = ">=4.41.0"
protobuf = "<4.21.3"
pandas = ">=1.1.4"
seaborn = ">=0.11.0"
wandb = "^0.15.12"
pillow = "^10.0.1"
gdown = "^4.7.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
